# swagger.yml
openapi: 3.0.3
info:
  title: custom-activity-service API
  version: 1.0.0
  description: >
    Service responsible for receiving execution events from Salesforce Marketing Cloud
    and orchestrating outbound operations against Genesys Cloud.
    This API is intended for internal integration use only.
servers:
  - url: https://{host}
    variables:
      host:
        default: custom-activity-service.example.com
security:
  - bearerAuth: []
...
paths:
  /execute:
    post:
      tags: [SFMC]
      summary: Receive execution event from SFMC
      description: >
        Endpoint invoked by Salesforce Marketing Cloud Journey Builder when a contact reaches
        the VoiceBot Custom Activity. This operation is idempotent based on requestId:
        duplicates are accepted but not reprocessed.
      operationId: execute
      parameters:
        - in: header
          name: X-Correlation-Id
          required: false
          schema:
            type: string
          description: Optional correlation identifier for tracing end-to-end requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VbEventIn"
            example:
              requestId: "REQ-0001"
              contactKey: "CK-0001"
              msisdn: "593999000001"
              contactListId: "331e970c-852c-4c37-903b-896dd1afacd3"
              campaignId: "e8ff3a36-7ae8-4508-9be1-2ba02c000000"
              status: "NEW"
              journeyId: "216b52d4-dbfc-4717-90dc-80a6a1fe4222"
              activityId: "629f5f20-f855-4525-9ed1-7e02496723a8"
      responses:
        "200":
          description: Event accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecuteResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    VbEventIn:
      type: object
      required:
        - requestId
        - contactKey
        - msisdn
        - contactListId
        - campaignId
        - status
        - journeyId
        - activityId
      properties:
        requestId:
          type: string
        contactKey:
          type: string
        msisdn:
          type: string
          description: "Phone number. Recommended format: 593 + 9 digits for Ecuador."
        contactListId:
          type: string
        campaignId:
          type: string
        status:
          type: string
        journeyId:
          type: string
        activityId:
          type: string
    ExecuteResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
          format: int64
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
...